transaction log

+++ 07-spring-transaction/src/main/java/com/albert/transaction/dao/BankDao.java
@@ -1,7 +1,7 @@
 package com.albert.transaction.dao;
 
 public interface BankDao {
-    void in(String name, int money);
+    void in(String name, long money);
 
-    void out(String name, int money);
+    void out(String name, long money);
 }
